prediction/prediction/backend/models/image_classifier.py
"""
Image Classifier using PyTorch ResNet18 Feature Extraction
Extracts deep features via pretrained ResNet18, classifies with a trained head.
Falls back to histogram features if torchvision is unavailable.
"""

import os
import io
import numpy as np
from PIL import Image
from sklearn.ensemble import GradientBoostingClassifier
import joblib

# Try to load torchvision for ResNet feature extraction
USE_RESNET = False
try:
    import torch
    import torchvision.transforms as transforms
    import torchvision.models as models
    USE_RESNET = True
    print("[OK] torchvision available - using ResNet18 feature extraction")
except ImportError:
    print("[FALLBACK] torchvision not available - falling back to histogram features")

LABELS = [
    "Benign Nevus",
    "Melanoma Suspect",
    "Dermatofibroma",
    "Basal Cell Carcinoma",
    "Actinic Keratosis",
    "Vascular Lesion"
]

LABEL_DETAILS = {
    "Benign Nevus": {
        "severity": "Low",
        "description": "Common mole, typically harmless. Regular visual monitoring mandatory.",
        "action": "CLINICAL DIRECTIVE: No immediate intervention. Implement monthly visual self-audit for ABCDE changes."
    },
    "Melanoma Suspect": {
        "severity": "Critical",
        "description": "Potential melanoma detected. High-risk malignant profiling.",
        "action": "URGENT PROTOCOL: Immediate dermatological excision and biopsy mandatory within 48 hours."
    },
    "Dermatofibroma": {
        "severity": "Low",
        "description": "Benign fibrous nodule. Confirmed harmless.",
        "action": "DIRECTIVE: No treatment required unless symptomatic. Elective excision optional."
    },
    "Basal Cell Carcinoma": {
        "severity": "Moderate",
        "description": "Persistent skin cancer detected. Slow-growing but requiring removal.",
        "action": "CLINICAL DIRECTIVE: Schedule surgical excision or cryotherapy within 14 days."
    },
    "Actinic Keratosis": {
        "severity": "Moderate",
        "description": "Pre-cancerous sun-induced lesion. High progression risk.",
        "action": "DIRECTIVE: Immediate intervention via cryotherapy or topical 5-FU protocol mandatory."
    },
    "Vascular Lesion": {
        "severity": "Low",
        "description": "Benign blood-vessel abnormality. Non-malignant.",
        "action": "PROTOCOL: Monitor for structural changes. Laser intervention elective for cosmetic improvement."
    }
}


class ImageClassifier:
    def __init__(self):
        self.model = None
        self.resnet = None
        self.transform = None
        self.saved_path = os.path.join(
            os.path.dirname(__file__), '..', 'models', 'saved', 'image_clf.pkl'
        )

        if USE_RESNET:
            self._init_resnet()

        self._ensure_model()

    def _init_resnet(self):
        """Initialize pretrained ResNet18 as a feature extractor (no grad)."""
        try:
            resnet = models.resnet18(weights=models.ResNet18_Weights.DEFAULT)
            # Remove final classification layer — use as feature extractor
            self.resnet = torch.nn.Sequential(*list(resnet.children())[:-1])
            self.resnet.eval()

            self.transform = transforms.Compose([
                transforms.Resize(256),
                transforms.CenterCrop(224),
                transforms.ToTensor(),
                transforms.Normalize(
                    mean=[0.485, 0.456, 0.406],
                    std=[0.229, 0.224, 0.225]
                )
            ])
            print("ResNet18 feature extractor loaded.")
        except Exception as e:
            print(f"ResNet init failed: {e}. Using histogram fallback.")
            self.resnet = None

    def _ensure_model(self):
        """Load saved classifier or create demo."""
        if os.path.exists(self.saved_path):
            self.model = joblib.load(self.saved_path)
        else:
            self._create_demo_model()

    def _create_demo_model(self):
        """Create a classifier trained on synthetic feature vectors."""
        np.random.seed(42)
        n_samples = 1200

        if self.resnet is not None:
            n_features = 512  # ResNet18 output
        else:
            n_features = 48 + 10  # histogram + texture stats

        X = np.random.randn(n_samples, n_features)
        y = np.random.randint(0, len(LABELS), n_samples)

        # Use GradientBoosting for better probability calibration
        self.model = GradientBoostingClassifier(
            n_estimators=150, max_depth=5, random_state=42
        )
        self.model.fit(X, y)

        os.makedirs(os.path.dirname(self.saved_path), exist_ok=True)
        joblib.dump(self.model, self.saved_path)
        print(f"Image classifier created ({n_features} features).")

    def extract_features(self, image_bytes: bytes) -> np.ndarray:
        """Extract features using ResNet18 or histogram fallback."""
        img = Image.open(io.BytesIO(image_bytes)).convert('RGB')

        if self.resnet is not None:
            return self._extract_resnet_features(img)
        else:
            return self._extract_histogram_features(img)

    def _extract_resnet_features(self, img: Image.Image) -> np.ndarray:
        """Extract 512-dim features from ResNet18's penultimate layer."""
        tensor = self.transform(img).unsqueeze(0)
        with torch.no_grad():
            features = self.resnet(tensor)
        return features.squeeze().numpy().reshape(1, -1)

    def _extract_histogram_features(self, img: Image.Image) -> np.ndarray:
        """Extract color histogram + texture statistics."""
        img = img.resize((224, 224))
        arr = np.array(img, dtype=float)

        features = []
        for ch in range(3):
            channel = arr[:, :, ch]
            # 16-bin histogram
            hist, _ = np.histogram(channel, bins=16, range=(0, 256))
            hist = hist / (hist.sum() + 1e-10)
            features.extend(hist)

        # Texture statistics (spatial variance, gradients)
        gray = np.mean(arr, axis=2)
        features.append(gray.mean() / 255.0)
        features.append(gray.std() / 255.0)
        features.append(np.median(gray) / 255.0)

        # Edge density (Sobel-like gradient magnitude)
        gx = np.diff(gray, axis=1)
        gy = np.diff(gray, axis=0)
        features.append(np.mean(np.abs(gx)))
        features.append(np.mean(np.abs(gy)))

        # Color ratios
        r, g, b = arr[:,:,0].mean(), arr[:,:,1].mean(), arr[:,:,2].mean()
        total = r + g + b + 1e-10
        features.extend([r/total, g/total, b/total])

        # Symmetry (compare left/right halves)
        mid = gray.shape[1] // 2
        left = gray[:, :mid]
        right = np.fliplr(gray[:, mid:mid+left.shape[1]])
        symmetry = 1.0 - np.mean(np.abs(left - right)) / 255.0
        features.append(symmetry)

        # Compactness
        features.append(arr.var() / (255.0 ** 2))

        return np.array(features).reshape(1, -1)

    def predict(self, image_bytes: bytes) -> dict:
        """Classify uploaded medical image."""
        features = self.extract_features(image_bytes)
        probabilities = self.model.predict_proba(features)[0]

        predicted_idx = int(np.argmax(probabilities))
        confidence = float(probabilities[predicted_idx])
        label = LABELS[predicted_idx]
        details = LABEL_DETAILS.get(label, {})

        top3_indices = np.argsort(probabilities)[::-1][:3]
        differential = [
            {
                "label": LABELS[i],
                "confidence": round(float(probabilities[i]) * 100, 1)
            }
            for i in top3_indices
        ]

        return {
            "classification": label,
            "confidence": round(confidence * 100, 1),
            "severity": details.get("severity", "Unknown"),
            "description": details.get("description", ""),
            "recommended_action": details.get("action", ""),
            "differential_diagnosis": differential,
            "feature_method": "ResNet18" if self.resnet else "Histogram+Texture"
        }
