prediction/prediction/backend/requirements.txt
Flask
Flask-CORS
scikit-learn
xgboost
lightgbm
pandas
numpy
spacy
transformers
torch
torchvision
shap
joblib
scipy
Pillow
lime
matplotlib
seaborn
pymongo
SpeechRecognition
pydub
easyocr
fuzzywuzzy
python-Levenshtein
librosa
soundfile
assemblyai
dnspython
